Implement a near compliant parser for media ranges and header parameters.
The quoted areas in the media ranges were being handled somewhat properly,
but we needed a more specialized form for other headers (notably, cookies).

librife introduces a fairly flexible parser and serializer for parameter
series with quoted text. It uses a much simpler algorithm than the implementation
used in libmedia which has been removed in favor of librife's.